% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Helper_Functions.R
\name{clean_text}
\alias{clean_text}
\title{Clean tweet text}
\usage{
clean_text(content)
}
\arguments{
\item{content}{A vector of strings (for instance get_pw_tweets()$content)}
}
\description{
Prepares tweets for text analysis by removing usernames (@user), URLs (http://...),
new lines, breaks and special symbols. Expands contractions (e.g. "won't" to "will not",
"I've" to "I have", etc.).
}
\examples{
\dontrun{

tweet.df <- get_pw_tweets(1, 5)
tweet.df$clean <- clean_text(tweet.df$content)

}
}
\author{
Katya Ognyanova \email{rstats@ognyanova.net}
}
\seealso{
\code{\link{get_pw_tweets}} \code{\link{get_pw_pol}}
          \code{\link{merge_pw}}
}

