% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Pw_Functions.R
\name{get_pw_pol}
\alias{get_pw_pol}
\title{Get politician data from the US Politwoops project}
\usage{
get_pw_pol(start.page = 1, end.page = "all")
}
\arguments{
\item{start.page}{The first page to get - defaults to page 1
(http://politwoops.sunlightfoundation.com/users?page=1)}

\item{end.page}{The last page to get - defaults to the last available page.}
}
\value{
Data frame with politician info from Politwoops.
}
\description{
This function scrapes politician data from the US Politwoops project
maintained by the Sunlight Foundation. The data is online at
\url{http://politwoops.sunlightfoundation.com/users}.
}
\examples{
\dontrun{

# Get all the politician data:
pol.df <- get_pw_pol()

# Get the first 5 pages of politicians:
pol.df <- get_pw_pol(1, 5)

}
}
\author{
Katya Ognyanova \email{rstats@ognyanova.net}
}
\seealso{
\code{\link{get_pw_tweets}}  \code{\link{merge_pw}}
}

