% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pw_Functions.R
\name{get_pw_tweets}
\alias{get_pw_tweets}
\title{Get deleted politician tweet data from the Politwoops project}
\usage{
get_pw_tweets(start.page = 1, end.page = "all", domain = NULL,
  json.ext = NULL)
}
\arguments{
\item{start.page}{The first json page to get - defaults to page 1.}

\item{end.page}{The last json page to get - defaults to the last available page.}

\item{domain}{The domain for the politwoops project of interest (do not include "http://").
Defaults to the US domain (politwoops.sunlightfoundation.com). For a list of other countries,
 see \url{http://politwoops.com/countries}.}

\item{json.ext}{The last part of the URL that points to the country JSON pages - defaults to 
"/index.json?page=" which seems to be correct for some countris (including the US and UK)
but not others. Future versions of this package will automate identifying this extension.}
}
\value{
Data frame with deleted tweet information from Politwoops.
}
\description{
This function downloads and combines json files from any of the countries
participating in the Politwoops project (defaults to the US).
}
\examples{
\dontrun{

# Get all the US Politwoops tweets:
tweet.df <- get_pw_tweets()

# Get the first 5 pages from the UK Politwoops tweets:
tweet.df <- get_pw_tweets(1, 5, "politwoops.co.uk")

# Get the first 5 pages from Argentina's Politwoops tweets:
tweet.df <- get_pw_tweets(1, 5, "www.politwoops.com/g/Argentina", ".json?page=")

}

}
\seealso{
\code{\link{get_pw_pol}} \code{\link{merge_pw}}
}
\author{
Katya Ognyanova \email{rstats@ognyanova.net}
}
