% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pw_Functions.R
\name{merge_pw}
\alias{merge_pw}
\title{Merge tweet and politician data}
\usage{
merge_pw(tweets, politicians)
}
\arguments{
\item{tweets}{Tweet data frame generated with  \code{\link{get_pw_tweets}}}

\item{politicians}{Politician data frame generated with  \code{\link{get_pw_pol}}}
}
\value{
Tweet data frame with added politician info for each tweet author.
}
\description{
Simple merge of tweets with politician data.
}
\examples{
\dontrun{

tweet.df <- get_pw_tweets(1, 5)
pol.df <-  get_pw_pol()
tweet.df <- merge_pw(tweet.df, pol.df)
 
 }
 
}
\seealso{
\code{\link{get_pw_tweets}} \code{\link{get_pw_pol}}
}
\author{
Katya Ognyanova \email{rstats@ognyanova.net}
}
