% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/PolitwoopsR.R
\docType{package}
\name{PolitwoopsR}
\alias{PolitwoopsR}
\alias{PolitwoopsR-package}
\title{Download Deleted Political Tweets From The Politwoops Project}
\description{
The PolitwoopsR package includes the functions get_pw_tweets
to extract tweet data from any Politwoops project, and get_pw_pol
to get a politician list from the US Politwoops project maintained by
the Sunlight Foundation. It also contains additional helper functions
for URL and text processing.
}
\author{
Katya Ognyanova \email{rstats@ognyanova.net}
}
\seealso{
\code{\link{get_pw_tweets}} \code{\link{get_pw_pol}}
            \code{\link{merge_pw}} \code{\link{url_domain}}
            \code{\link{url_expand}} \code{\link{clean_text}}
}

