% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Helper_Functions.R
\name{url_expand}
\alias{url_expand}
\title{Expand shortened URLs in tweets}
\usage{
url_expand(tweets)
}
\arguments{
\item{tweets}{A data frame of tweet info created with \code{\link{get_pw_tweets}} or
\code{\link{merge_pw}}}
}
\value{
A vector of expanded URLs.
}
\description{
Expand the shortened URLs in a deleted tweet data frame. This may take some time
since every tweet has to go through the longurl.org API. Note also that many
of deleted tweets have broken or wrong links - those of course cannot be expanded.
}
\examples{
\dontrun{

tweet.df <- get_pw_tweets(1, 5)
tweet.df$url.exp <- url_expand(tweet.df)

 }
}
\author{
Katya Ognyanova \email{rstats@ognyanova.net}
}
\seealso{
\code{\link{get_pw_tweets}} \code{\link{get_pw_pol}}
          \code{\link{merge_pw}} \code{\link{url_domain}}
}

