% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Pw_Functions.R
\name{get_pw_tweets}
\alias{get_pw_tweets}
\title{Get deleted politician tweet data from the Politwoops project}
\usage{
get_pw_tweets(start.page = 1, end.page = "all", domain = NULL)
}
\arguments{
\item{start.page}{The first json page to get - defaults to page 1.}

\item{end.page}{The last json page to get - defaults to the last available page.}

\item{domain}{The domain for the politwoops project of interest (do not include "http://").
Defaults to the US domain (politwoops.sunlightfoundation.com). For a list of other countries,
 see \url{http://politwoops.com/countries}.}
}
\value{
Data frame with deleted tweet information from Politwoops.
}
\description{
This function downloads and combines json files from any of the countries
participating in the Politwoops project (defaults to the US).
}
\examples{
\dontrun{

# Get all the US Politwoops tweets:
tweet.df <- get_pw_tweets()

# Get the first 5 pages from the UK Politwoops tweets:
tweet.df <- get_pw_tweets(1, 5, "politwoops.co.uk")
}
}
\author{
Katya Ognyanova \email{rstats@ognyanova.net}
}
\seealso{
\code{\link{get_pw_pol}} \code{\link{merge_pw}}
}

